Modificación de A entero

+++ ejercicioGrupalSprintBoot/src/main/java/com/adt/ejerciciogrupalsprintboot/repositories/IUsuariosRepository.java
@@ -4,5 +4,5 @@ package com.adt.ejerciciogrupalsprintboot.repositories;
 import com.adt.ejerciciogrupalsprintboot.models.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUsuariosRepository extends JpaRepository<Usuarios, Long> {
+public interface IUsuariosRepository extends JpaRepository<Usuarios, Integer> {
 }
